MLOPS/mlflow-project/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bleach==6.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.102.0
docker==7.1.0
exceptiongroup==1.3.1
fastapi==0.110.3
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
greenlet==3.3.2
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaggle==1.6.17
kiwisolver==1.5.0
Mako==1.3.10
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.1
mlflow==2.17.2
mlflow-skinny==2.17.2
mlflow-tracing==3.10.1
numpy==2.2.6
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==24.2
pandas==2.2.3
pillow==12.1.1
prettytable==3.17.0
protobuf==5.29.6
pyarrow==17.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2026.1.post1
pywin32==311
PyYAML==6.0.3
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.37.2
text-unidecode==1.3
threadpoolctl==3.6.0
tomli==2.4.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.30.6
waitress==3.0.2
wcwidth==0.6.0
webencodings==0.5.1
Werkzeug==3.1.7
zipp==3.23.0
